chore: rename HeaderV3 to Header
