+++ frontend/kesaseteli/employer/src/components/header/Header.tsx
@@ -4,7 +4,7 @@ import useUserQuery from 'kesaseteli/employer/hooks/backend/useUserQuery';
 import { useRouter } from 'next/router';
 import { useTranslation } from 'next-i18next';
 import React from 'react';
-import BaseHeader from 'shared/components/header/HeaderV3';
+import BaseHeader from 'shared/components/header/Header';
 import { Language, SUPPORTED_LANGUAGES } from 'shared/i18n/i18n';
 import { OptionType } from 'shared/types/common';
